разделила код
1. Обновила импорты в файле auth-modal.js, pages.js. 2. Слушатель Марины перенесла в новый файл removeFromFav.js - Марине написала об этом. 3. Проверила у себя после разделения - все работает. 4. В файле card.js оставила вынесеный код закоментированым, на случай если по сле риквеста в dev что-то сломается. Если всё норм, обновлюсь ещё раз и удалю ненужные коменты.
