+++ src/js/favorites/remove-favorite.js
@@ -19,9 +19,9 @@ async function onRemoveFavoriteItem(e) {
 }
 
 export async function delFavItem(e) {
-  if (!e.target.parentNode.classList.contains('card__favorite-btn--orange')) {
-    return;
-  }
+  // if (!e.target.parentNode.classList.contains('card__favorite-btn--orange')) {
+  //   return;
+  // }
   const cardId = getId(e);
   const userToken = getUserToken();
